backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
scikit-learn==1.4.2
fairlearn==0.10.0
openai
python-dotenv
loguru
requests
